Move util files to util package
Also rename `incident/utils.py` to `util/mapping_utils.py`.
